fix(gui): bump gradio to >=5.20 + ruff to >=0.9.3 so the GUI boots
Local ``./scripts/try-it.sh all`` hit two blockers after torch installed
cleanly on Py 3.13 Mac arm64:

1. ``TypeError: argument of type 'bool' is not iterable`` from
   ``gradio_client==1.6.0`` — known bug hit by pydantic inputs that
   emit ``additionalProperties: True`` in their JSON schema.
   Fixed in gradio 5.20+ / gradio-client 1.7+.
2. Stale ``.venv`` leftover from switching between ``[dev]`` /
   ``[train]`` / ``[gui]`` in separate ``uv sync`` calls: gradio
   got half-uninstalled, leaving an empty namespace-package stub
   that ``pytest.importorskip("gradio")`` happily passed through.

Fixes
-----
- Bump ``gradio>=5.20,<6`` + ``gradio-client>=1.7,<2`` in ``[gui]``.
- Bump ``ruff>=0.9.3,<1.0`` in ``[dev]`` — gradio 5.20 pulls ruff
  0.9.3 as a build-time dep, and the older 0.8 pin conflicted.
- Pass ``show_api=False`` in ``launch_app`` so Gradio skips the
  JS-client schema introspection path (it's not used by the tabs).
- Harden ``test_build_app_returns_blocks`` gate to detect a stale
  gradio stub via ``importlib.util.find_spec("gradio.blocks")``.
- Combine the ``[train]`` + ``[gui]`` ``uv sync`` calls in
  ``scripts/try-it.sh`` so switching modes never leaves the venv in
  a half-uninstalled state.
- Fix ruff 0.9.3 / 0.15 lint regressions: drop redundant ``int(round(...))``
  casts, escape regex metacharacters in ``pytest.raises(match=)``,
  and collapse a ``x if cond else x`` ternary.
- Silence pyright ``reportPrivateImportUsage`` for torch's non-
  exported public factories (``torch.tensor``, ``torch.from_numpy``,
  ``torch.arange``, ``torch.zeros_like``).
- Drop the removed ``local_dir_use_symlinks`` kwarg from
  ``huggingface_hub.snapshot_download`` (deleted in hub 1.0).
- Fix two gradio tab ``build()`` return annotations (``-> Any``).

Verified locally on macOS arm64 + Python 3.13: ``uv sync --extra
dev --extra train --extra gui`` resolves cleanly to ``torch==2.11``,
``gradio==5.50``, ``gradio-client==1.14``; ``uv run pytest`` prints
``310 passed, 2 skipped``; a scripted ``app.launch`` probes
``/config`` and gets a clean 200 response.

+++ src/openpathai/explain/attention_rollout.py
@@ -87,7 +87,7 @@ def rollout_from_matrices(
     cls_row = rollout[0, 1:] if has_cls_token else np.mean(rollout, axis=0)
 
     num_patches = cls_row.shape[0]
-    side = int(round(np.sqrt(num_patches)))
+    side = round(np.sqrt(num_patches))
     if side * side != num_patches:
         raise ValueError(f"attention_rollout expects a square patch grid; got {num_patches} tokens")
     return cls_row.reshape(side, side)
